+++ CHANGES.rst
@@ -24,6 +24,9 @@ in progress
   and ``RACKER_WDM_MACHINE``.
 - Add environment variable ``RACKER_WDM_PROVIDER`` to reconfigure the
   Vagrant virtualization backend differently than VirtualBox.
+- Documentation: Add use case how to build a Python package within a
+  Windows environment, using Microsoft Visual C++ Build Tools 2015 and
+  Anaconda, both installed using Chocolatey, and ``cibuildwheel``.
 
 
 2022-05-20 0.2.0